requirements.txt
fastapi
uvicorn
streamlit
python-multipart
pymupdf
langchain
langchain-community
langchain_openai
faiss-cpu
openai
tiktoken
python-dotenv
